Sandbox: Limit the memory buffers of the Sandbox IP to 32 bits.
BUG=123875
TEST=sbox_uniitests
Review URL: https://chromiumcodereview.appspot.com/10369002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@136987 0039d316-1c4b-4281-b951-d872f2087c98

+++ sandbox/src/internal_types.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2006-2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -28,9 +28,9 @@ enum ArgType {
 // Encapsulates a pointer to a buffer and the size of the buffer.
 class CountedBuffer {
  public:
-  CountedBuffer(void* buffer, size_t size) : size_(size), buffer_(buffer) {}
+  CountedBuffer(void* buffer, uint32 size) : size_(size), buffer_(buffer) {}
 
-  size_t Size() const {
+  uint32 Size() const {
     return size_;
   }
 
@@ -39,7 +39,7 @@ class CountedBuffer {
   }
 
  private:
-  size_t size_;
+  uint32 size_;
   void* buffer_;
 };
